% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/document_test.R
\name{add}
\alias{add}
\title{Add two numbers}
\usage{
add(x, y)
}
\arguments{
\item{x}{A number}

\item{y}{Another number}
}
\value{
The sum of x and y
}
\description{
This function adds two numbers together.
}
